♻️ Separate legacy curators from new curators
